flasky/requirements.txt
Flask
Flask-Bootstrap
Flask-Moment
Flask-WTF
python-dotenv
requests
WTForms
email-validator
